requirements/base.txt
Django==4.1.4
django-dotenv==1.4.1
django-debug-toolbar>=3.2,<4
django-cors-headers==3.13.0
wagtail==4.2
wagtailfontawesome>=1.1.3,<1.2
wagtail-simple-gallery>=0.9.0
wagtail-headless-preview>=0.5.0
psycopg2>=2.7,<3.0

django-crispy-forms==1.14.0
wagtail-django-recaptcha==1.0
wagtail-metadata==4.0.2
Pygments==2.14.0

django-countries==7.5
django-allauth==0.52.0
django-widget-tweaks==1.4.12

crispy-bootstrap5==0.7
django-comments-xtd==2.9.9
django-contrib-comments==2.2.0
crispy-tailwind==0.5.0
emoji==2.2.0





